chore(faas): fix cycle dependency
